+++ Mechaduino/analogFastWrite.h
@@ -9,7 +9,7 @@
    \brief SAMD products have only one reference for ADC
 */
 
-extern void digitalWriteDirect(int PIN, bool val);
+extern void digitalWriteDirect(byte PIN, bool val);
 
 extern void analogFastWrite( uint32_t ulPin, uint32_t ulValue ) ;
